增加武纺C语言课程部分第三章编程作业

+++ OJ-2-1.c
@@ -16,7 +16,7 @@ int main()
 {
     int a, b, c;
     scanf("%d %d %d", &a, &b, &c);
-    printf("%d %.1f", a + b + c, (float)(a+b+c)/3.0);
+    printf("%d %.1f", a + b + c, (a+b+c)/3.0);
 
 
     return 0;
